perf: Remote-friendlier central directory search
The stock `archive/zip` package first looks for the "end of central
directory" marker in the last 1K, then in the last 65K. I'm not sure
where exactly that optimization makes sense (lots of extremely small
zip files?), but it usually means twice the network requests when
used with an httpkit/htfs.File.

So, we just do the last 65K now.